optimize workspace structure
